# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TaskFlow API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workspaces:
    get:
      operationId: listWorkspaces
      tags: [workspaces]
      summary: List workspaces for the current user
      responses:
        "200":
          description: List of workspaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workspace"
    post:
      operationId: createWorkspace
      tags: [workspaces]
      summary: Create a workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceInput"
      responses:
        "201":
          description: Created workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
  /workspaces/{workspaceId}:
    get:
      operationId: getWorkspace
      tags: [workspaces]
      summary: Get a workspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: Workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
    patch:
      operationId: updateWorkspace
      tags: [workspaces]
      summary: Update a workspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceUpdate"
      responses:
        "200":
          description: Updated workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
    delete:
      operationId: deleteWorkspace
      tags: [workspaces]
      summary: Delete a workspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "204":
          description: Deleted
...
  /workspaces/{workspaceId}/tasks:
    get:
      operationId: listWorkspaceTasks
      tags: [tasks]
      summary: List tasks across a workspace, with optional filters
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - name: projectId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/TaskStatus"
        - name: priority
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/TaskPriority"
        - name: assigneeId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
  /tasks:
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{taskId}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - $ref: "#/components/parameters/TaskId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "204":
          description: Deleted
...
components:
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        type: integer
...
    TaskId:
      name: taskId
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    WorkspaceType:
      type: string
      enum: [personal, team]
    WorkspaceRole:
      type: string
      enum: [owner, admin, member]
...
    TaskStatus:
      type: string
      enum: [todo, in_progress, done]
    TaskPriority:
      type: string
      enum: [low, medium, high, urgent]
...
    Workspace:
      type: object
      required: [id, name, type, role, memberCount, projectCount, taskCount, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        type: { $ref: "#/components/schemas/WorkspaceType" }
        role: { $ref: "#/components/schemas/WorkspaceRole" }
        memberCount: { type: integer }
        projectCount: { type: integer }
        taskCount: { type: integer }
        createdAt: { type: string, format: date-time }
    WorkspaceInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        type: { $ref: "#/components/schemas/WorkspaceType" }
    WorkspaceUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
...
    Task:
      type: object
      required: [id, projectId, projectName, workspaceId, title, description, status, priority, dueDate, assigneeId, createdAt, updatedAt]
      properties:
        id: { type: integer }
        projectId: { type: integer }
        projectName: { type: string }
        workspaceId: { type: integer }
        title: { type: string }
        description: { type: ["string", "null"] }
        status: { $ref: "#/components/schemas/TaskStatus" }
        priority: { $ref: "#/components/schemas/TaskPriority" }
        dueDate: { type: ["string", "null"], format: date }
        assigneeId: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    TaskInput:
      type: object
      required: [projectId, title]
      properties:
        projectId: { type: integer }
        title: { type: string, minLength: 1 }
        description: { type: string }
        status: { $ref: "#/components/schemas/TaskStatus" }
        priority: { $ref: "#/components/schemas/TaskPriority" }
        dueDate: { type: ["string", "null"], format: date }
        assigneeId: { type: ["string", "null"] }
    TaskUpdate:
      type: object
      properties:
        projectId: { type: integer }
        title: { type: string, minLength: 1 }
        description: { type: ["string", "null"] }
        status: { $ref: "#/components/schemas/TaskStatus" }
        priority: { $ref: "#/components/schemas/TaskPriority" }
        dueDate: { type: ["string", "null"], format: date }
        assigneeId: { type: ["string", "null"] }
...