# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * TaskFlow API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TaskPriority } from './taskPriority';
import type { TaskStatus } from './taskStatus';

export interface Task {
  id: number;
  projectId: number;
  projectName: string;
  workspaceId: number;
  title: string;
  /** @nullable */
  description: string | null;
  status: TaskStatus;
  priority: TaskPriority;
  /** @nullable */
  dueDate: Date | null;
  /** @nullable */
  assigneeId: string | null;
  createdAt: Date;
  updatedAt: Date;
}