# lib/api-zod/src/generated/types/taskInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * TaskFlow API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TaskPriority } from './taskPriority';
import type { TaskStatus } from './taskStatus';

export interface TaskInput {
  projectId: number;
  /** @minLength 1 */
  title: string;
  description?: string;
  status?: TaskStatus;
  priority?: TaskPriority;
  /** @nullable */
  dueDate?: Date | null;
  /** @nullable */
  assigneeId?: string | null;
}