# lib/api-zod/src/generated/types/taskStatus.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * TaskFlow API specification
 * OpenAPI spec version: 0.1.0
 */

export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];


export const TaskStatus = {
  todo: 'todo',
  in_progress: 'in_progress',
  done: 'done',
} as const;