# lib/api-zod/src/generated/types/workspace.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * TaskFlow API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WorkspaceRole } from './workspaceRole';
import type { WorkspaceType } from './workspaceType';

export interface Workspace {
  id: number;
  name: string;
  type: WorkspaceType;
  role: WorkspaceRole;
  memberCount: number;
  projectCount: number;
  taskCount: number;
  createdAt: Date;
}