# lib/api-zod/src/generated/types/workspaceType.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * TaskFlow API specification
 * OpenAPI spec version: 0.1.0
 */

export type WorkspaceType = typeof WorkspaceType[keyof typeof WorkspaceType];


export const WorkspaceType = {
  personal: 'personal',
  team: 'team',
} as const;