# lib/api-zod/src/generated/types/workspaceUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * TaskFlow API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WorkspaceUpdate {
  /** @minLength 1 */
  name?: string;
}